Add Self-serve method & registry module (#746)
We want people to fully onboard without intervention, this PR adds the
first registry module to do so. This registry module can call `owner()`
or `getCCIPAdmin` on a token contract, which will set the Admin to the
result of that call, if the msg.sender is equal to the result of the
call. This means any `owner` of an existing token will be able to claim
Admin privileges on the AdminRegistry through the registration module.


## Review scope

Please review the entire TokenAdminRegistry, not just the changed parts