Return dest denomimination token amount from pool (#762)
This PR enables the destination chain pools to specify the amount that
was minted or released from the pool. Using this functionality, it is
possible to create token pools that support tokens with different
decimals on different chains. There are various ways to implement this,
for instance, using the token pool return data to encode the source
chain decimals or some shared decimal schema. A pool could also maintain
a full decimal mapping of all remote chains and do the calculations on
the remote chain.

We have also removed the duplicated check for the contract size of a
pool. Since `_callWithExactGasSafeReturnData` already does this check,
we only have to catch it in `_trialExecute` to safely remove the
duplicated check.


Due to the optimization in the check, the PR reduces gas usage while
adding the token pool functionality