[TT-1106] Fix parsing of versions from ecr for compatibility testing (#765)
## Motivation
The logic to pull the latest image version from ecr was not working
consistently

## Solution
Break out the logic into a function with a bit more documentation for
what is happening. Allow to pass a regex in for specific differences in
image tagging practices and use the linux sort function for the sorting.