Median liquidity per chain fault tolerance validation (#797)
## Motivation

Currently, we don’t count values of the reported liquidity in each
observation to ensure we have at least f+1 values for each chain, which
means that we fail the case of empty values (e.g. an oracle that fails
to get the values).

## Solution

Count liquidities reported for each chain and ensure the count exceeds
the minimum fault tolerance threshold (f+1).