Make plugin robust against outcome manipulation (#799)
## Motivation

The OCR plugin fails to deduplicate the values of slices in each
oracle’s reported observation. This allows a single node to introduce
duplicate values in its observation slices, causing the plugin to
erroneously trust that certain values are the consensus among the DON.

## Solution

Extending observation validation to include deduplication validation.