Self serve migration (#796)
This PR will add two new pools, one burnMint and one lockRelease. These
will be used as proxies for the existing <1.5 token pools when the lanes
upgrade to 1.5. Please see
`contracts/src/v0.8/ccip/test/legacy/BurnMintTokenPoolAndProxy.t.sol`
for a full step-by-step guide on how this upgrade works for both the
1.0/1.2 and the 1.4 pools.


Pools deployed after self serve is live will NOT use these proxies, but
rather the clean versions that do not implement the proxy behaviour.
Because of this, we only need proxies for the pools that are currently
live: burnMint and LockRelease. USDC will cleanly upgrade by deploying
new pools and doesn't require the proxy. All other pool flavours are not
live at the moment

---------

Co-authored-by: Justin Kaseman <justinkaseman@live.com>