use real events in tests (#925)
Since we use 0.8.24, we can now import events from other contracts. This
PR replaces almost all copied events in test with references to the real
events. Using this method means we ensure events in the real contracts
and the tests stay in sync, while reducing the need for manual updates.
It also reduces the size of test files


We also remove the CustomTokenPool since it is not used

We removed the last instanced of `/// @notice <function>` which have
been a strange leftover from the days the testsuite was still hardhat
based.