Add check for existence of ECR images used as test dependencies in E2E EVM compatibility tests (#13271)
* Add check for existence of ECR images used as test dependencies

The code changes include adding a new step in the workflow file `evm-version-compatibility-tests.yml` to check if the images used as test dependencies exist in the ECR. This step is triggered when there is a change in the test dependency or when the workflow is manually dispatched. The step checks if the latest Ethereum client image exists in the ECR and fails the workflow if it doesn't exist.

* Add CTF workflow trigger job

* Add permissions for the trigger job

* Add actions: read permission

* Refactor workflow

* Fix

* fix

* Use update-internal-mirrors action from CTF to wait until image is updated

* Bump action version

* Add image check to evm compatibility tests

* Bump action version

* Check images for other clients

* Check only ethereum/client-go in evm-version-compatibility-tests.yml