Adding setters to CommitStore for GasEstimator and SourceMaxGasPrice (#1047)
## Motivation

CommitStore currently is constructed with dependencies from both a
Source Chain Relayer and Dest Chain Relayer. This violates a key
constraint of our system when moving to LOOPPs. This change is one step
to separate the construction of a `CommitStore` from requiring the
source max gas price and source gas price estimator.
## Solution

Instead of setting `sourceMaxGasPrice` and `estimator` in the
construction of a `CommitStore`, we now use explicit setter methods.

This will be consumed in future PRs to add a provider based
implementation of a gas estimator in the reporting plugin factory.

---------

Co-authored-by: dimitris <dimitrios.kouveris@smartcontract.com>