[CCIP-2639] ccip capability launcher component (#1123)
## Motivation

We need a component that consumes capability registry syncer updates and
launches CCIP oracles accordingly.

## Solution

Build a launcher component that implements the `registrysyncer.Launcher`
interface. The setup of the `launcher` package is as follows:

* bluegreen.go: defines structures that encapsulates ccip deployments,
i.e exec and commit oracles, and the blue/green deployment logic.
* diff.go: pure functions that diff registry states in order to
determine if there are any added, updated, or deleted dons that we
should care about.
* launcher.go: defines the `launcher` type which implements the
`registrysyncer.Launcher` interface. The `Launch` function is called by
the registry syncer whenever a new registry state is fetched. Note that
this registry state is not always different, we will most likely be
called with the same state more often than not, as the registry doesn't
change that often. It uses a handful of pure functions and applies state
updates only in a single function, `processDiff`.

On the contracts:
* Break out the types defined in `CCIPConfig.sol` into a separate
CCIPConfigTypes.sol which is then imported in CCIPConfig.sol and the OCR
config encoding interface, a helper used in tests.
* Use `Internal.OCRPluginType` instead of `PluginType` in order to be
consistent with the OCR3Base contracts that will be deployed on dests