Integration test Setup Nodes and Add DON [CCIP-2810] (#1190)
Setting up a basic test that have basic OCR3 Nodes. Each Node can access
3 EVM chains that are created and connected to each others. Check
[PR](https://github.com/smartcontractkit/ccip/pull/1148) for more
details on creating the chains and connecting them.

As JobSpecs for CCIP are not ready. The main goal of this PR is to have
the nodes and AddDon on capability registry for these nodes for each
chain. Once JobSpec is ready we'll be able to test that the nodes picks
up the changes and deploys the plugins, start sending, receiving,..etc.