Don't propagate parent's context for the background replay (#1219)
## Solution

It adds the missing part to this PR
https://github.com/smartcontractkit/ccip/pull/1201

We can't pass the foreground routine context to the task run in the
background and could take minutes or hours to finish. Instead of that,
we create a background context and expose cancelFn. This is exactly how
it used to work before introducing loopp layer