Avoid sending already verified signatures when transmitting (#1224)
## Motivation

Exec plugin submits reports which do not have their signatures checked
on chain. This is to save gas, because signature checking is expensive
and executions can only be for committed data which is already
authenticated. It is executed using a OCR2 plugin as it handles the turn
taking functionality.

Gas can be saved by not even submitting the signatures as calldata. Each
signature is about 64 bytes and it costs 16 per non-zero byte of
calldata.

## Solution
In the Exec Provider, configure the Transmitter to skip sending the
signatures.
The exec provider was modified in
https://github.com/smartcontractkit/chainlink/pull/13761