Rename PriceRegistry -> FeeQuoter (#1350)
Rename the PriceRegistry to FeeQuoter. 

This PR restores the 1.2 PriceRegistry interface which is still used for
CCIP 1.5. The FeeQuoter is a superset of that interface, which is why
all tests can use the feeQuoter, even in 1.5.


This PR does NOT claim to rename all offchain instances. It only handles
onchain and wrapper/mock based references