Multi-Mechanism USDC Token Pools (#1280)
## Motivation

Currently all USDC token transfers occur using Burn/Mint with calls to
Circle's CCTP system for cross-chain transfers. Certain chains in CCIP
have non-canonical USDC deployed and want to integrate the ability to
use CCIP as a canonical bridge whereby canonical-USDC is locked on the
source-chain and then minted on the destination chain. Then at a later
date, when circle decides to add that chain to CCTP, the non-canonical
USDC will be converted into canonical by burning the locked-tokens on
the source-chain. To accomplish this, the token pool needs to be capable
of identifying when tokens should be burned/minted with CCTP, or if they
should use regular lock-release, and act accordingly.

## Solution

`USDCTokenPool.sol` was been modified into a new file to allow for two
different mechanisms to be used simultaneously. The primary mechanism,
which is opt-out, and the secondary-mechanism which is opt in. These
mechanisms are configured on a per-chain-selector basis, and must be
manually enabled. In this implementation CCTP is the primary mechanism,
and Lock/Release is the alternative mechanism.

There are two new files
1. `HybridLockReleaseUSDCTokenPool.sol`, The actual implementation of
the hybrid pool structure.
3. `USDCBridgeMigrator.sol`, which contains the logic necessary to
conform to [Circle's migration policy
guide](https://github.com/circlefin/stablecoin-evm/blob/master/doc/bridged_USDC_standard.md#transferring-the-roles-to-circle)