feat: CCIP-3020 Add AllowList Config struct and implementation  (#1335)
The key objective of allowlisting senders is to restrict the senders
(EOA or smart contract) that can send CCIP Messages from a given
blockchain to a specified destination blockchain.

While acknowledging this additional check may come with gas cost, we aim
to achieve this objective with minimal gas cost increments.

The CCIP OnRamp contract must be able to allow only specific senders
(EOA or smart contract), identified by address, to send CCIP messages to
a specific destination blockchain, identified by
destinationChainSelector.

CCIP messages for a specific destination blockchain for which the sender
is not included in the allowlist must be rejected. A proper error must
be returned.

It should be possible to activate/deactivate the allowlist capability of
an OnRamp for a given destination chain. Options include:

- A flag that enables/disables allowlisting capability
- The selected approach should consider gas cost, clarity and
simplicity.

Only a designed entity must be able to update the allowlist
configuration. By default this is CCIP Owner, but it must be possible
for CCIP Owner to also add a third party (i.e. AllowListAdmin) to have
this privilege.

- Add new properties to DestChainConfig Struct
```js
  /// @dev Struct to hold the configs for a destination chain
  /// @dev sequenceNumber, allowListEnabled, router will all be packed in 1 slot
  struct DestChainConfig {
    // The last used sequence number. This is zero in the case where no messages have yet been sent.
    // 0 is not a valid sequence number for any real transaction.
    uint64 sequenceNumber;
    // boolean indicator to specify if allowList check is enabled
    bool allowListEnabled;
    // This is the local router address that is allowed to send messages to the destination chain.
    // This is NOT the receiving router address on the destination chain.
    IRouter router;
    // This is the list of addresses allowed to send messages from onRamp
    EnumerableSet.AddressSet allowedSendersList;
  }
```

- functions to applyAllowList

```js
function applyAllowListUpdates(AllowListConfigArgs[] calldata allowListConfigArgsItems) external {
```

- a struct to hold the applyAllowList Information

```js
  struct AllowListConfigArgs {
    uint64 destChainSelector; // Destination chain selector
    bool allowListEnabled; // indicator to check if the allowedSendersList is enabled
    address[] addedAllowlistedSenders; // list of senders to be added to the allowedSendersList
    address[] removedAllowlistedSenders; // list of senders to be removed from the allowedSendersList
  }
```

---------

Co-authored-by: app-token-issuer-infra-releng[bot] <120227048+app-token-issuer-infra-releng[bot]@users.noreply.github.com>