[CCIP-3360] contracts/src/v0.8/ccip: merge latest changes from the ccip repo (#14345)
* Misc golfs and fixes (#1402)

Co-authored-by: app-token-issuer-infra-releng[bot] <120227048+app-token-issuer-infra-releng[bot]@users.noreply.github.com>

* Increase max signers (#1405)

Improve tests and error checks

---------

Co-authored-by: app-token-issuer-infra-releng[bot] <120227048+app-token-issuer-infra-releng[bot]@users.noreply.github.com>

* rename CCIPSendRequested to CCIPMessageSent (#1409)

* Skip report execution on curse (#1408)

## Motivation

The goal of this PR is to remove reverts on lane cursing for DON
execution transactions. If a lane is cursed, reverting will cause the
whole execution transaction to fail so any execution reports for other
lanes will be blocked.

## Solution

For DON execution transactions the behavior is now to skip the report
and emit a `SkippedReportExecution`. For manual execution we still
revert, otherwise the transaction will silently fail for the users.

---------

Co-authored-by: app-token-issuer-infra-releng[bot] <120227048+app-token-issuer-infra-releng[bot]@users.noreply.github.com>

* remove bootstrapP2PIds (#1388)

## Motivation

Bootstrap P2P IDs are no longer needed in the OCR config in CCIPConfig.
See #1348's description for more details.

## Solution

Remove bootstrap P2P IDs from the OCR config in CCIPConfig.

## Related PRs
https://github.com/smartcontractkit/chainlink-ccip/pull/89

* fixes

* Integrate RMNRemote and OffRamp (#1360)

Remove the requirement for self-transmitted RMN blessings

Allow the commit DON to include RMN blessings at commitment time
This PR has a dependency on changes to
[chainlink-ccip](https://github.com/smartcontractkit/chainlink-ccip/pull/84)

---------

Co-authored-by: Makram <makramkd@users.noreply.github.com>

* contracts/scripts: reduce offramp optimizations (#1414)

## Motivation

The offramp was failing to deploy due to too many optimizations causing
a max code size exceeded error.

## Solution

Reduce the number of optimizations by 1.

Also, add a test that deploys v1.6 contracts and ensures that they are
deployable on the simulated backend.

* more fixes

* remove isBlessed from contract reader cfg

* deploy RMNRemote instead of MockRMN

* bump chainlink-ccip

* bump chainlink-ccip and add changesets

---------

Co-authored-by: Rens Rooimans <github@rensrooimans.nl>
Co-authored-by: app-token-issuer-infra-releng[bot] <120227048+app-token-issuer-infra-releng[bot]@users.noreply.github.com>
Co-authored-by: Suryansh <39276431+0xsuryansh@users.noreply.github.com>
Co-authored-by: Ryan <80392855+RayXpub@users.noreply.github.com>
Co-authored-by: Ryan Hall <RyanRHall@users.noreply.github.com>