feat: add getter to return all source chain config (#1477)
## Motivation
The current offchain initialization process involves calling
getSourceChainConfig(sourceChainSelector), which retrieves
configurations for a specific source chain selector. However, there's no
mechanism to fetch a complete list of supported source chains during the
initialization phase. This limitation complicates the process, as the
full list of source chain selectors is not readily available.

## Solution
To address this, the proposed solution introduces an enumerable set of
all source chain selectors. By implementing a getAllSourceChainConfig()
function, we can easily retrieve the complete list of source chains
on-chain. This eliminates the need for additional rounds of OCR gossip
to share known source chains, simplifying initialization and improving
efficiency.

---------

Signed-off-by: 0xsuryansh <suryansh.shrivastava@smartcontract.com>
Co-authored-by: app-token-issuer-infra-releng[bot] <120227048+app-token-issuer-infra-releng[bot]@users.noreply.github.com>
Co-authored-by: Rens Rooimans <github@rensrooimans.nl>