feat(chain): return network data
Currently for graphql query chains and chain, they do not support returning network/chain type such as evm, solana, or aptos etc.

For `chain` and `chains` graphql query, this commit add support for returning the network associated with the chain.

This can then be used on the operator UI to show a list of supported chains instead of just hardcoded to EVM.

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1150