[DPA-1157]: feat: introduce cosmosKeys & starknetKeys graphql query  (#14764)
* feat(cosmos): introduce cosmosKeys graphql query

Similar to eTHKeys and solanaKeys, we need cosmosKeys that will be used by the operator ui to show a list of addresses.

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1157

* feat(starknet): introduce starknetKeys graphql query

Similar to eTHKeys and solanaKeys, we need starknetKeys that will be used by the operator ui to show a list of addresses.

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1157