Initial move of integration-tests/deployment to /deployment as a separate module.
This updates the dependencies, and makes the integration-tests depend on both chainlink and the deployments package.

A few things to fix.  The go.mod entry for deployments is off, because there is no version currently checked-in. It's overridden in a replace statement, but I'm not sure if there are any contexts in which that goes wrong. It just seems weird to keep an incorrect version number there.

Also:
  * Rename the node client to nodeclient, to disambiguate.
  * Move the various non-product elements into 'environment'