fix(operator-ui): show chain type next to keyBundleId
For the ocr2 keybundle id select box, it lists all the keys based on the ID and it's hard to tell which one we want. Adding the key type (aptos, evm etc) to the values in the selectbox, something like "0xaaabbbccc (aptos)" so it can be easily identified now that we support more than EVM chain.

Context: https://github.com/smartcontractkit/operator-ui/pull/96

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1321