Contract Reader Handle Empty Bytes
In some cases, a contract call will return an empty result `0x`, which decodes to an empty set of bytes. The codec
can't decode this empty set except in specific cases where a slice of bytes is the expected result.

This commit adds a short-circuit to decoding an empty result where the return value is unaltered.