An initial cut of a GetState mechanism for the changeset/migrations system of deployment.
The basic API is GetState(env, StateKey[MyStateStruct]{}) which returns MyStateStruct. Under the hood, there is a stateManager on the Environment, which handles the backend. In chainlink/deployment (and integration-tests) this will use an in-memory state store, which will be primed with data for the test. In real deployments, a different backend will pull from json or backend information platform protos (depending on the evolution of that system). From the point of view of the ChangeSet execution function, this will be invisible. The ChangeSet should just be able to call the GetState function and the appropriate system will return the struct, or an error.

A separate function is used, rather than a method on Environment or on the state manager, because methods cannot have type paramters other than those of their owned type, and this function needs to handle any type coercion internally. Otherwise we'd do env.state.GetState(key). Sadly, this limitation of go generics means we have to use a detached outer function.