fix(ccip): readjust error check
In the previous [PR](https://github.com/smartcontractkit/chainlink/pull/16458/files#diff-f6eeb14804b5d16769abc3f48ad18187965263d5e5d29345e4079e85199de731R127), there was a bug that was introduced that made the code to exit early and did not fall in to the `ConfirmIfNoErrorWithABI` so better error message can be decoded when mcms is not enabled. This commit brings back that behaviour.