Send Metering Report to Beholder
This PR collects metering information from workflow steps, builds a report and sends it as a custom message to beholder.
Metering values are persisted along with the workflow.

The metering report is currently sent as a string value for `BaseMessage` as a temporary measure, but the proto definition
exists such that the metering report can be sent as a custom proto message when the infrastructure exists to do so.