feat(operations): introduce idempotency behaviour
When a sequence or an operation is executed, it nows checks for previous execution in the reporter and if a successful execution was found in the past, it will skip the current execution and return the previous successful result.

Execution will not be skipped if the previous execution was a failure.

An operation or sequence uniqueness is determine by its definition (id + version +description) and its input

The current execution report will not include any reports for operations/sequences that were skipped.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-21