fix(operations): make report fully serialiazable
Currently operations report has 2 fields which are not fully serializable
- Timestamp time.Time
- Err error

`Timestamp` is not serialisable because it has exported fields (potential data lost), however by changing it to pointer, it implements json.marshaler and json.unmarshaler which formats time to RFC 3339.

`Error` is an interface, and error object does not have an exported field so serialising error becomes an empty struct {}. Creating a new error with exported `Message` field allows serialization to write the message to json.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-50