Send and Emit Metering Report
This commit adds support for sending a metering report to a billing service. It is expected that the billing client
will not be set in production until a complete billing client is made available in the future. In this case, the report
is sent to the billing client if it is set.

Additionally, a metering report is not always sent to the beholder service. A global beholder client is expected to be
set and a custom metering report message is supplied to the client.