fix(operations): introduce util function to convert to generic type
In some cases, we want to return the report created by the operations or sequence to the changeset level, however those level deals with Report[any,any], this commit introduce function to help with the conversion without the user having to convert themselves. We are required to perform an explicit conversion because golang does not support Covariance concept.