refactor: remove redundant ApplyChangesets
With the removal of the legacy MCMS, the 2 methods ApplyChangesets and ApplyChangesetsV2 are now very similar with the exception of the timelock contract field that is no longer used. This commit migrate all usages to a single ApplyChangesets method and clean up all the timelockcontract usage that is no longer used.