core/capabilities/ccip: add hashing fuzz tests
Add fuzz tests for the ccipevm offchain message hasher. I ran the fuzzer
locally for 2 minutes and it generated 115 entries in the corpus.
I could run it for longer to get more data points but this seems like
a good start.

Added the corpus to the testdata/fuzz/FuzzMessageHasher directory, that
way all these entries will be used as regression tests when `go test`
runs in CI.